Update files, fix lints
