+++ 3rdparty/py/requirements-all.txt
@@ -1,5 +1,6 @@
 boto3==1.34.37
 celery==5.3.6
+charset-normalizer==3.3.2
 dj-rest-auth[with_social]==5.0.2
 Django==3.2.23
 django-cors-headers==4.3.1
@@ -13,7 +14,7 @@ google-auth-oauthlib==1.2.0
 greenlet==3.0.3
 gunicorn[gevent, setproctitle]==21.2.0
 redis==5.0.1
-requests==2.27.0
+requests==2.31.0
 sentry-sdk==1.40.2
 tablib[xlsx]==3.5.0
 tweepy==4.14.0
