+++ pesacheck_meedan_bridge/py/BUILD
@@ -4,6 +4,7 @@ python_sources(
     "3rdparty/py:requirements-all#requests",
     "3rdparty/py:requirements-all#sentry-sdk",
     "3rdparty/py:requirements-all#environs",
+    "3rdparty/py:requirements-all#charset-normalizer",
   ]
 )
 
@@ -12,7 +13,7 @@ pex_binary(
     dependencies=[
         ":lib",
     ],
-    entry_point="service.py",
+    entry_point="main.py",
     include_sources=False,
     include_tools=True,
     layout="packed",
@@ -23,7 +24,7 @@ pex_binary(
     dependencies=[
         ":lib",
     ],
-    entry_point="service.py",
+    entry_point="main.py",
     include_requirements=False,
     include_tools=True,
     layout="packed",
